+++ sql/schema/v0.sql
@@ -1070,6 +1070,8 @@ CREATE TABLE
         CONSTRAINT "WorkflowTriggers_pkey" PRIMARY KEY ("id")
     );
 
+CREATE TYPE idempotency_method AS ENUM ('TTL', 'STATUS');
+
 -- CreateTable
 CREATE TABLE
     "WorkflowVersion" (
@@ -1090,6 +1092,7 @@ CREATE TABLE
         "inputJsonSchema" JSONB,
         "idempotencyKeyExpression" TEXT,
         "idempotencyKeyTtlMs" BIGINT,
+        "idempotencyMethod" idempotency_method,
         CONSTRAINT "WorkflowVersion_pkey" PRIMARY KEY ("id")
     );